# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingCurrency =
  (typeof ListingCurrency)[keyof typeof ListingCurrency];

export const ListingCurrency = {
  DZD: "DZD",
} as const;

export type ListingType = (typeof ListingType)[keyof typeof ListingType];

export const ListingType = {
  sell: "sell",
  rent: "rent",
} as const;

export type ListingCondition =
  (typeof ListingCondition)[keyof typeof ListingCondition];

export const ListingCondition = {
  new: "new",
  good: "good",
  used: "used",
  needs_repair: "needs_repair",
} as const;

export type ListingSellerType =
  (typeof ListingSellerType)[keyof typeof ListingSellerType];

export const ListingSellerType = {
  individual: "individual",
  agency: "agency",
} as const;

export interface Listing {
  id: string;
  title: string;
  description: string;
  price: number;
  currency: ListingCurrency;
  type: ListingType;
  categoryId: string;
  categoryName: string;
  categoryIcon?: string;
  wilayaId: string;
  wilayaName: string;
  city: string;
  address?: string;
  images: string[];
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  floor?: number;
  totalFloors?: number;
  furnished?: boolean;
  condition?: ListingCondition;
  sellerName: string;
  sellerPhone: string;
  sellerAvatar?: string;
  sellerType: ListingSellerType;
  agencyName?: string;
  featured: boolean;
  verified: boolean;
  views: number;
  createdAt: string;
  updatedAt?: string;
  amenities?: string[];
  tags?: string[];
}

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  limit: number;
  offset: number;
}

export type CreateListingInputType =
  (typeof CreateListingInputType)[keyof typeof CreateListingInputType];

export const CreateListingInputType = {
  sell: "sell",
  rent: "rent",
} as const;

export type CreateListingInputCondition =
  (typeof CreateListingInputCondition)[keyof typeof CreateListingInputCondition];

export const CreateListingInputCondition = {
  new: "new",
  good: "good",
  used: "used",
  needs_repair: "needs_repair",
} as const;

export type CreateListingInputSellerType =
  (typeof CreateListingInputSellerType)[keyof typeof CreateListingInputSellerType];

export const CreateListingInputSellerType = {
  individual: "individual",
  agency: "agency",
} as const;

export interface CreateListingInput {
  title: string;
  description: string;
  price: number;
  type: CreateListingInputType;
  categoryId: string;
  wilayaId: string;
  city: string;
  address?: string;
  images?: string[];
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  furnished?: boolean;
  condition?: CreateListingInputCondition;
  sellerName: string;
  sellerPhone: string;
  sellerType: CreateListingInputSellerType;
  agencyName?: string;
  amenities?: string[];
}

export interface Category {
  id: string;
  name: string;
  nameAr: string;
  icon: string;
  description?: string;
  listingCount: number;
}

export type WilayaRegion = (typeof WilayaRegion)[keyof typeof WilayaRegion];

export const WilayaRegion = {
  north: "north",
  center: "center",
  east: "east",
  west: "west",
  south: "south",
} as const;

export interface Wilaya {
  id: string;
  code: number;
  name: string;
  nameAr: string;
  region: WilayaRegion;
  listingCount: number;
}

export interface ContactSellerInput {
  name: string;
  phone: string;
  email?: string;
  message: string;
}

export interface ContactSellerResponse {
  success: boolean;
  message: string;
}

export interface MarketplaceStats {
  totalListings: number;
  totalSellers: number;
  wilayas: number;
  categories: number;
  newToday: number;
}

export type GetListingsParams = {
  categoryId?: string;
  wilayaId?: string;
  type?: GetListingsType;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type GetListingsType =
  (typeof GetListingsType)[keyof typeof GetListingsType];

export const GetListingsType = {
  sell: "sell",
  rent: "rent",
} as const;