# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Darkom - Home & Furniture Marketplace Algeria API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: listings
    description: Listing operations
  - name: categories
    description: Category operations
  - name: wilayas
    description: Wilaya (province) operations
  - name: messages
    description: Message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: wilayaId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sell, rent]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

  /listings/{id}:
    get:
      operationId: getListingById
      tags: [listings]
      summary: Get a single listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Listing not found

  /listings/{id}/contact:
    post:
      operationId: contactSeller
      tags: [listings]
      summary: Send a message to the seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSellerInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSellerResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /wilayas:
    get:
      operationId: getWilayas
      tags: [wilayas]
      summary: Get all Algerian wilayas
      responses:
        "200":
          description: List of wilayas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wilaya"

  /stats:
    get:
      operationId: getStats
      tags: [listings]
      summary: Get marketplace statistics
      responses:
        "200":
          description: Marketplace stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Listing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
          enum: [DZD]
        type:
          type: string
          enum: [sell, rent]
        categoryId:
          type: string
        categoryName:
          type: string
        categoryIcon:
          type: string
        wilayaId:
          type: string
        wilayaName:
          type: string
        city:
          type: string
        address:
          type: string
        images:
          type: array
          items:
            type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        area:
          type: number
        floor:
          type: integer
        totalFloors:
          type: integer
        furnished:
          type: boolean
        condition:
          type: string
          enum: [new, good, used, needs_repair]
        sellerName:
          type: string
        sellerPhone:
          type: string
        sellerAvatar:
          type: string
        sellerType:
          type: string
          enum: [individual, agency]
        agencyName:
          type: string
        featured:
          type: boolean
        verified:
          type: boolean
        views:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        amenities:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - price
        - currency
        - type
        - categoryId
        - categoryName
        - wilayaId
        - wilayaName
        - city
        - images
        - sellerName
        - sellerPhone
        - sellerType
        - featured
        - verified
        - views
        - createdAt

    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - listings
        - total
        - limit
        - offset

    CreateListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        type:
          type: string
          enum: [sell, rent]
        categoryId:
          type: string
        wilayaId:
          type: string
        city:
          type: string
        address:
          type: string
        images:
          type: array
          items:
            type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        area:
          type: number
        furnished:
          type: boolean
        condition:
          type: string
          enum: [new, good, used, needs_repair]
        sellerName:
          type: string
        sellerPhone:
          type: string
        sellerType:
          type: string
          enum: [individual, agency]
        agencyName:
          type: string
        amenities:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - price
        - type
        - categoryId
        - wilayaId
        - city
        - sellerName
        - sellerPhone
        - sellerType

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        description:
          type: string
        listingCount:
          type: integer
      required:
        - id
        - name
        - nameAr
        - icon
        - listingCount

    Wilaya:
      type: object
      properties:
        id:
          type: string
        code:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        region:
          type: string
          enum: [north, center, east, west, south]
        listingCount:
          type: integer
      required:
        - id
        - code
        - name
        - nameAr
        - region
        - listingCount

    ContactSellerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        message:
          type: string
      required:
        - name
        - phone
        - message

    ContactSellerResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    MarketplaceStats:
      type: object
      properties:
        totalListings:
          type: integer
        totalSellers:
          type: integer
        wilayas:
          type: integer
        categories:
          type: integer
        newToday:
          type: integer
      required:
        - totalListings
        - totalSellers
        - wilayas
        - categories
        - newToday