# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all listings with optional filters
 */
export const getListingsQueryLimitDefault = 20;
export const getListingsQueryOffsetDefault = 0;

export const GetListingsQueryParams = zod.object({
  categoryId: zod.coerce.string().optional(),
  wilayaId: zod.coerce.string().optional(),
  type: zod.enum(["sell", "rent"]).optional(),
  minPrice: zod.coerce.number().optional(),
  maxPrice: zod.coerce.number().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.boolean().optional(),
  limit: zod.coerce.number().default(getListingsQueryLimitDefault),
  offset: zod.coerce.number().default(getListingsQueryOffsetDefault),
});

export const GetListingsResponse = zod.object({
  listings: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string(),
      price: zod.number(),
      currency: zod.enum(["DZD"]),
      type: zod.enum(["sell", "rent"]),
      categoryId: zod.string(),
      categoryName: zod.string(),
      categoryIcon: zod.string().optional(),
      wilayaId: zod.string(),
      wilayaName: zod.string(),
      city: zod.string(),
      address: zod.string().optional(),
      images: zod.array(zod.string()),
      bedrooms: zod.number().optional(),
      bathrooms: zod.number().optional(),
      area: zod.number().optional(),
      floor: zod.number().optional(),
      totalFloors: zod.number().optional(),
      furnished: zod.boolean().optional(),
      condition: zod.enum(["new", "good", "used", "needs_repair"]).optional(),
      sellerName: zod.string(),
      sellerPhone: zod.string(),
      sellerAvatar: zod.string().optional(),
      sellerType: zod.enum(["individual", "agency"]),
      agencyName: zod.string().optional(),
      featured: zod.boolean(),
      verified: zod.boolean(),
      views: zod.number(),
      createdAt: zod.string(),
      updatedAt: zod.string().optional(),
      amenities: zod.array(zod.string()).optional(),
      tags: zod.array(zod.string()).optional(),
    }),
  ),
  total: zod.number(),
  limit: zod.number(),
  offset: zod.number(),
});

/**
 * @summary Create a new listing
 */
export const CreateListingBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  type: zod.enum(["sell", "rent"]),
  categoryId: zod.string(),
  wilayaId: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  images: zod.array(zod.string()).optional(),
  bedrooms: zod.number().optional(),
  bathrooms: zod.number().optional(),
  area: zod.number().optional(),
  furnished: zod.boolean().optional(),
  condition: zod.enum(["new", "good", "used", "needs_repair"]).optional(),
  sellerName: zod.string(),
  sellerPhone: zod.string(),
  sellerType: zod.enum(["individual", "agency"]),
  agencyName: zod.string().optional(),
  amenities: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a single listing by ID
 */
export const GetListingByIdParams = zod.object({
  id: zod.coerce.string(),
});

export const GetListingByIdResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string(),
  price: zod.number(),
  currency: zod.enum(["DZD"]),
  type: zod.enum(["sell", "rent"]),
  categoryId: zod.string(),
  categoryName: zod.string(),
  categoryIcon: zod.string().optional(),
  wilayaId: zod.string(),
  wilayaName: zod.string(),
  city: zod.string(),
  address: zod.string().optional(),
  images: zod.array(zod.string()),
  bedrooms: zod.number().optional(),
  bathrooms: zod.number().optional(),
  area: zod.number().optional(),
  floor: zod.number().optional(),
  totalFloors: zod.number().optional(),
  furnished: zod.boolean().optional(),
  condition: zod.enum(["new", "good", "used", "needs_repair"]).optional(),
  sellerName: zod.string(),
  sellerPhone: zod.string(),
  sellerAvatar: zod.string().optional(),
  sellerType: zod.enum(["individual", "agency"]),
  agencyName: zod.string().optional(),
  featured: zod.boolean(),
  verified: zod.boolean(),
  views: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string().optional(),
  amenities: zod.array(zod.string()).optional(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Send a message to the seller
 */
export const ContactSellerParams = zod.object({
  id: zod.coerce.string(),
});

export const ContactSellerBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  message: zod.string(),
});

export const ContactSellerResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all categories
 */
export const GetCategoriesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  nameAr: zod.string(),
  icon: zod.string(),
  description: zod.string().optional(),
  listingCount: zod.number(),
});
export const GetCategoriesResponse = zod.array(GetCategoriesResponseItem);

/**
 * @summary Get all Algerian wilayas
 */
export const GetWilayasResponseItem = zod.object({
  id: zod.string(),
  code: zod.number(),
  name: zod.string(),
  nameAr: zod.string(),
  region: zod.enum(["north", "center", "east", "west", "south"]),
  listingCount: zod.number(),
});
export const GetWilayasResponse = zod.array(GetWilayasResponseItem);

/**
 * @summary Get marketplace statistics
 */
export const GetStatsResponse = zod.object({
  totalListings: zod.number(),
  totalSellers: zod.number(),
  wilayas: zod.number(),
  categories: zod.number(),
  newToday: zod.number(),
});