# lib/api-zod/src/generated/types/contactSellerResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSellerResponse {
  success: boolean;
  message: string;
}