# lib/api-zod/src/generated/types/createListingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateListingInputCondition } from "./createListingInputCondition";
import type { CreateListingInputSellerType } from "./createListingInputSellerType";
import type { CreateListingInputType } from "./createListingInputType";

export interface CreateListingInput {
  title: string;
  description: string;
  price: number;
  type: CreateListingInputType;
  categoryId: string;
  wilayaId: string;
  city: string;
  address?: string;
  images?: string[];
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  furnished?: boolean;
  condition?: CreateListingInputCondition;
  sellerName: string;
  sellerPhone: string;
  sellerType: CreateListingInputSellerType;
  agencyName?: string;
  amenities?: string[];
}