# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingCondition } from "./listingCondition";
import type { ListingCurrency } from "./listingCurrency";
import type { ListingSellerType } from "./listingSellerType";
import type { ListingType } from "./listingType";

export interface Listing {
  id: string;
  title: string;
  description: string;
  price: number;
  currency: ListingCurrency;
  type: ListingType;
  categoryId: string;
  categoryName: string;
  categoryIcon?: string;
  wilayaId: string;
  wilayaName: string;
  city: string;
  address?: string;
  images: string[];
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  floor?: number;
  totalFloors?: number;
  furnished?: boolean;
  condition?: ListingCondition;
  sellerName: string;
  sellerPhone: string;
  sellerAvatar?: string;
  sellerType: ListingSellerType;
  agencyName?: string;
  featured: boolean;
  verified: boolean;
  views: number;
  createdAt: string;
  updatedAt?: string;
  amenities?: string[];
  tags?: string[];
}