# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceStats {
  totalListings: number;
  totalSellers: number;
  wilayas: number;
  categories: number;
  newToday: number;
}