# lib/api-zod/src/generated/types/wilaya.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darkom - Home & Furniture Marketplace Algeria API
 * OpenAPI spec version: 0.1.0
 */
import type { WilayaRegion } from "./wilayaRegion";

export interface Wilaya {
  id: string;
  code: number;
  name: string;
  nameAr: string;
  region: WilayaRegion;
  listingCount: number;
}